+++ lua/obsidian/command.lua
@@ -5,25 +5,6 @@ local echo = require "obsidian.echo"
 local util = require "obsidian.util"
 
 local command = {}
-local attach_find_command = function(client, opts)
-  -- sort_by is:
-  -- "none"
-  -- "path"
-  -- "modified"
-  -- "accessed"
-  -- "created"
-  local sort_by = client.opts.sort_by
-  if sort_by == nil then
-    return opts
-  end
-  local sort_reversed = client.opts.sort_reversed
-  local sort = "sortr" -- default sort is reverse
-  if sort_reversed == false then
-    sort = "sort"
-  end
-  opts.find_command = { "rg", "--files", "--" .. sort, sort_by }
-  return opts
-end
 
 ---Check the directory for notes with missing/invalid frontmatter.
 ---
@@ -336,8 +317,8 @@ command.template = function(client, data)
             end)
             return true
           end,
+          find_command = util.build_find_cmd(".", client.opts.sort_by, client.opts.sort_reversed),
         }
-        opts = attach_find_command(client, opts)
         require("telescope.builtin").find_files(opts)
       end
       choose_template()
@@ -348,10 +329,9 @@ command.template = function(client, data)
       if not has_fzf_lua then
         util.implementation_unavailable()
       end
-      local cmd = vim.tbl_flatten { util.FIND_CMD, { ".", "-name", "'*.md'" } }
-      cmd = util.table_params_to_str(cmd)
+      local cmd = util.build_find_cmd(".", client.opts.sort_by, client.opts.sort_reversed)
       fzf_lua.files {
-        cmd = cmd,
+        cmd = util.table_params_to_str(cmd),
         cwd = tostring(templates_dir),
         file_icons = false,
         actions = {
@@ -376,9 +356,8 @@ command.template = function(client, data)
           vim.api.nvim_del_user_command "ApplyTemplate"
         end, { nargs = 1, bang = true })
 
-        local base_cmd = vim.tbl_flatten { util.FIND_CMD, { tostring(templates_dir), "-name", "'*.md'" } }
-        base_cmd = util.table_params_to_str(base_cmd)
-        local fzf_options = { source = base_cmd, sink = "ApplyTemplate" }
+        local cmd = util.build_find_cmd(tostring(templates_dir), client.opts.sort_by, client.opts.sort_reversed)
+        local fzf_options = { source = util.table_params_to_str(cmd), sink = "ApplyTemplate" }
         vim.api.nvim_call_function("fzf#run", {
           vim.api.nvim_call_function("fzf#wrap", { fzf_options }),
         })
@@ -393,8 +372,8 @@ end
 ---Quick switch to an obsidian note
 ---
 ---@param client obsidian.Client
----@param data table
-command.quick_switch = function(client, data)
+---@param _ table
+command.quick_switch = function(client, _)
   local dir = tostring(client.dir)
 
   client:_run_with_finder_backend(":ObsidianQuickSwitch", {
@@ -405,8 +384,11 @@ command.quick_switch = function(client, data)
         util.implementation_unavailable()
       end
       -- Search with telescope.nvim
-      local opts = attach_find_command(client, { cwd = dir, search_file = "*.md" })
-      telescope.find_files(opts)
+      telescope.find_files {
+        cwd = dir,
+        search_file = "*.md",
+        find_command = util.build_find_cmd(".", client.opts.sort_by, client.opts.sort_reversed),
+      }
     end,
     ["fzf-lua"] = function()
       local has_fzf_lua, fzf_lua = pcall(require, "fzf-lua")
@@ -414,16 +396,14 @@ command.quick_switch = function(client, data)
       if not has_fzf_lua then
         util.implementation_unavailable()
       end
-      local cmd = vim.tbl_flatten { util.FIND_CMD, { ".", "-name", "'*.md'" } }
-      cmd = util.table_params_to_str(cmd)
-      fzf_lua.files { cmd = cmd, cwd = tostring(client.dir) }
+      local cmd = util.build_find_cmd(".", client.opts.sort_by, client.opts.sort_reversed)
+      fzf_lua.files { cmd = util.table_params_to_str(cmd), cwd = tostring(client.dir) }
     end,
     ["fzf.vim"] = function()
       -- Fall back to trying with fzf.vim
       local has_fzf, _ = pcall(function()
-        local base_cmd = vim.tbl_flatten { util.FIND_CMD, { dir, "-name", "'*.md'" } }
-        base_cmd = util.table_params_to_str(base_cmd)
-        local fzf_options = { source = base_cmd, sink = "e" }
+        local cmd = util.build_find_cmd(dir, client.opts.sort_by, client.opts.sort_reversed)
+        local fzf_options = { source = util.table_params_to_str(cmd), sink = "e" }
         vim.api.nvim_call_function("fzf#run", {
           vim.api.nvim_call_function("fzf#wrap", { fzf_options }),
         })